+++ docs/html/search/all_10.js
@@ -1,5 +1,6 @@
 var searchData=
 [
-  ['temperaturewriter_78',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html',1,'ace_segment::TemperatureWriter'],['../classace__segment_1_1TemperatureWriter.html#a7ff33a2cf83fcfbf411d048d5b522982',1,'ace_segment::TemperatureWriter::TemperatureWriter()']]],
-  ['tm1637module_79',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment::Tm1637Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module::Tm1637Module()']]]
+  ['temperaturewriter_82',['TemperatureWriter',['../classace__segment_1_1TemperatureWriter.html',1,'ace_segment::TemperatureWriter'],['../classace__segment_1_1TemperatureWriter.html#a3d0dc943b6b9dbb9ab43ac1720189bd1',1,'ace_segment::TemperatureWriter::TemperatureWriter()']]],
+  ['tm1637module_83',['Tm1637Module',['../classace__segment_1_1Tm1637Module.html',1,'ace_segment::Tm1637Module&lt; T_TMII, T_DIGITS &gt;'],['../classace__segment_1_1Tm1637Module.html#a9cc401ceeed8292b26a93d5dcc08312e',1,'ace_segment::Tm1637Module::Tm1637Module()']]],
+  ['twowireinterface_84',['TwoWireInterface',['../classace__segment_1_1TwoWireInterface.html',1,'ace_segment']]]
 ];