+++ docs/html/search/all_c.js
@@ -1,4 +1,4 @@
 var searchData=
 [
-  ['numberwriter_54',['NumberWriter',['../classace__segment_1_1NumberWriter.html',1,'ace_segment::NumberWriter'],['../classace__segment_1_1NumberWriter.html#afd63f269907c5dee9fb6a41a00eb3d37',1,'ace_segment::NumberWriter::NumberWriter()']]]
+  ['numberwriter_56',['NumberWriter',['../classace__segment_1_1NumberWriter.html',1,'ace_segment::NumberWriter'],['../classace__segment_1_1NumberWriter.html#abb61529de2cbf876061915895d68f01d',1,'ace_segment::NumberWriter::NumberWriter()']]]
 ];