+++ docs/html/search/classes_3.js
@@ -1,7 +1,8 @@
 var searchData=
 [
-  ['hardspifastinterface_110',['HardSpiFastInterface',['../classace__segment_1_1HardSpiFastInterface.html',1,'ace_segment']]],
-  ['hardspiinterface_111',['HardSpiInterface',['../classace__segment_1_1HardSpiInterface.html',1,'ace_segment']]],
-  ['hc595module_112',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
-  ['hybridmodule_113',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
+  ['hardspifastinterface_116',['HardSpiFastInterface',['../classace__segment_1_1HardSpiFastInterface.html',1,'ace_segment']]],
+  ['hardspiinterface_117',['HardSpiInterface',['../classace__segment_1_1HardSpiInterface.html',1,'ace_segment']]],
+  ['hc595module_118',['Hc595Module',['../classace__segment_1_1Hc595Module.html',1,'ace_segment']]],
+  ['ht16k33module_119',['Ht16k33Module',['../classace__segment_1_1Ht16k33Module.html',1,'ace_segment']]],
+  ['hybridmodule_120',['HybridModule',['../classace__segment_1_1HybridModule.html',1,'ace_segment']]]
 ];