+++ docs/html/search/classes_4.js
@@ -1,12 +1,11 @@
 var searchData=
 [
-  ['leddisplay_114',['LedDisplay',['../classace__segment_1_1LedDisplay.html',1,'ace_segment']]],
-  ['ledmatrixbase_115',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
-  ['ledmatrixdirect_116',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdirect_3c_20gpiointerface_20_3e_117',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
-  ['ledmatrixdualhc595_118',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_119',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_120',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
-  ['ledmodule_121',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]],
-  ['levelwriter_122',['LevelWriter',['../classace__segment_1_1LevelWriter.html',1,'ace_segment']]]
+  ['ledmatrixbase_121',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html',1,'ace_segment']]],
+  ['ledmatrixdirect_122',['LedMatrixDirect',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdirect_3c_20gpiointerface_20_3e_123',['LedMatrixDirect&lt; GpioInterface &gt;',['../classace__segment_1_1LedMatrixDirect.html',1,'ace_segment']]],
+  ['ledmatrixdualhc595_124',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_125',['LedMatrixSingleHc595',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmatrixsinglehc595_3c_20t_5fspii_2c_20gpiointerface_20_3e_126',['LedMatrixSingleHc595&lt; T_SPII, GpioInterface &gt;',['../classace__segment_1_1LedMatrixSingleHc595.html',1,'ace_segment']]],
+  ['ledmodule_127',['LedModule',['../classace__segment_1_1LedModule.html',1,'ace_segment']]],
+  ['levelwriter_128',['LevelWriter',['../classace__segment_1_1LevelWriter.html',1,'ace_segment']]]
 ];