+++ docs/html/search/functions_8.js
@@ -1,8 +1,7 @@
 var searchData=
 [
-  ['leddisplay_161',['LedDisplay',['../classace__segment_1_1LedDisplay.html#a6cda5f98bdfbc8dd1661dc41e1e94d02',1,'ace_segment::LedDisplay']]],
-  ['ledmatrixbase_162',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html#ae46efc84d412faff4617fe461b912131',1,'ace_segment::LedMatrixBase']]],
-  ['ledmatrixdualhc595_163',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html#aa32dd16e69cd42930243c083ee4dc287',1,'ace_segment::LedMatrixDualHc595']]],
-  ['ledmodule_164',['LedModule',['../classace__segment_1_1LedModule.html#a5941f3f6bcee5c757cb3983a4542b61f',1,'ace_segment::LedModule']]],
-  ['levelwriter_165',['LevelWriter',['../classace__segment_1_1LevelWriter.html#a262b1aa9d2f28c8c82eb96f52f438189',1,'ace_segment::LevelWriter']]]
+  ['ledmatrixbase_173',['LedMatrixBase',['../classace__segment_1_1LedMatrixBase.html#ae46efc84d412faff4617fe461b912131',1,'ace_segment::LedMatrixBase']]],
+  ['ledmatrixdualhc595_174',['LedMatrixDualHc595',['../classace__segment_1_1LedMatrixDualHc595.html#aa32dd16e69cd42930243c083ee4dc287',1,'ace_segment::LedMatrixDualHc595']]],
+  ['ledmodule_175',['LedModule',['../classace__segment_1_1LedModule.html#a5941f3f6bcee5c757cb3983a4542b61f',1,'ace_segment::LedModule::LedModule()'],['../classace__segment_1_1CharWriter.html#a44d4a19a3deee0f1ac1783dd2d610efc',1,'ace_segment::CharWriter::ledModule()'],['../classace__segment_1_1ClockWriter.html#a5ebd90e614a3cb6b41625634cf13aea2',1,'ace_segment::ClockWriter::ledModule()'],['../classace__segment_1_1LevelWriter.html#a4284ff0ce92453182596db0c273881b5',1,'ace_segment::LevelWriter::ledModule()'],['../classace__segment_1_1NumberWriter.html#a9be2815ba8f1bfd219c37989005910a0',1,'ace_segment::NumberWriter::ledModule()'],['../classace__segment_1_1PatternWriter.html#af3d95175f0490a375a3d0171f3e51511',1,'ace_segment::PatternWriter::ledModule()'],['../classace__segment_1_1StringScroller.html#ab81eb92243e2bfbdf9d445e7dfc41578',1,'ace_segment::StringScroller::ledModule()'],['../classace__segment_1_1StringWriter.html#a8ebe52a50690e195e94895adbca57616',1,'ace_segment::StringWriter::ledModule()'],['../classace__segment_1_1TemperatureWriter.html#a505122cb61cbe740291351210e86123a',1,'ace_segment::TemperatureWriter::ledModule()']]],
+  ['levelwriter_176',['LevelWriter',['../classace__segment_1_1LevelWriter.html#a9472935c5224a65be532ed6ab191ca37',1,'ace_segment::LevelWriter']]]
 ];