+++ docs/html/search/variables_0.js
@@ -1,11 +1,11 @@
 var searchData=
 [
-  ['kcharminus_213',['kCharMinus',['../classace__segment_1_1ClockWriter.html#ab535943b0a616a41bf686b74f9a9fc31',1,'ace_segment::ClockWriter::kCharMinus()'],['../classace__segment_1_1NumberWriter.html#a039ac6a2528399ac73b384f6227d3e6a',1,'ace_segment::NumberWriter::kCharMinus()']]],
-  ['kcharspace_214',['kCharSpace',['../classace__segment_1_1ClockWriter.html#a5432977b4275f45adcc3764296a7c8ce',1,'ace_segment::ClockWriter::kCharSpace()'],['../classace__segment_1_1NumberWriter.html#ad5907a93c3f37b3cef81d3427688d18c',1,'ace_segment::NumberWriter::kCharSpace()']]],
-  ['knumhexchars_215',['kNumHexChars',['../classace__segment_1_1NumberWriter.html#a2da066eed77d440c8d2df88ad38407bb',1,'ace_segment::NumberWriter']]],
-  ['kpatterna_216',['kPatternA',['../classace__segment_1_1ClockWriter.html#a53759004793ba6bcc90408725829434f',1,'ace_segment::ClockWriter']]],
-  ['kpatternc_217',['kPatternC',['../classace__segment_1_1TemperatureWriter.html#ab708932843ab2f48fe22f1a95a1ea156',1,'ace_segment::TemperatureWriter']]],
-  ['kpatterndegree_218',['kPatternDegree',['../classace__segment_1_1TemperatureWriter.html#a1be9f982f9466b3afa0d01051dc04855',1,'ace_segment::TemperatureWriter']]],
-  ['kpatternf_219',['kPatternF',['../classace__segment_1_1TemperatureWriter.html#aab41ee81d8fe308c9d0c0d9c1e3c381d',1,'ace_segment::TemperatureWriter']]],
-  ['kpatternp_220',['kPatternP',['../classace__segment_1_1ClockWriter.html#aba763a1a69889af49dc7a34c972a309b',1,'ace_segment::ClockWriter']]]
+  ['kcharminus_228',['kCharMinus',['../classace__segment_1_1ClockWriter.html#ab535943b0a616a41bf686b74f9a9fc31',1,'ace_segment::ClockWriter::kCharMinus()'],['../classace__segment_1_1NumberWriter.html#a039ac6a2528399ac73b384f6227d3e6a',1,'ace_segment::NumberWriter::kCharMinus()']]],
+  ['kcharspace_229',['kCharSpace',['../classace__segment_1_1ClockWriter.html#a5432977b4275f45adcc3764296a7c8ce',1,'ace_segment::ClockWriter::kCharSpace()'],['../classace__segment_1_1NumberWriter.html#ad5907a93c3f37b3cef81d3427688d18c',1,'ace_segment::NumberWriter::kCharSpace()']]],
+  ['knumhexchars_230',['kNumHexChars',['../classace__segment_1_1NumberWriter.html#a2da066eed77d440c8d2df88ad38407bb',1,'ace_segment::NumberWriter']]],
+  ['kpatterna_231',['kPatternA',['../classace__segment_1_1ClockWriter.html#a53759004793ba6bcc90408725829434f',1,'ace_segment::ClockWriter']]],
+  ['kpatternc_232',['kPatternC',['../classace__segment_1_1TemperatureWriter.html#ab708932843ab2f48fe22f1a95a1ea156',1,'ace_segment::TemperatureWriter']]],
+  ['kpatterndegree_233',['kPatternDegree',['../classace__segment_1_1TemperatureWriter.html#a1be9f982f9466b3afa0d01051dc04855',1,'ace_segment::TemperatureWriter']]],
+  ['kpatternf_234',['kPatternF',['../classace__segment_1_1TemperatureWriter.html#aab41ee81d8fe308c9d0c0d9c1e3c381d',1,'ace_segment::TemperatureWriter']]],
+  ['kpatternp_235',['kPatternP',['../classace__segment_1_1ClockWriter.html#aba763a1a69889af49dc7a34c972a309b',1,'ace_segment::ClockWriter']]]
 ];