+++ CHANGELOG.md
@@ -1,6 +1,39 @@
 # Changelog
 
 * Unreleased
+* 0.7 (2021-06-16)
+    * Support HT16K33 LED modules
+        * Add `Ht16k33Module` class to support 4-digit LED display from
+          Adafruit or one of its clones.
+        * Add `TwoWireInterface` as a thin abstraction between `TwoWire` class
+          and the `Ht16k33Module`, to avoid including `<Wire.h>` which pulls in
+          about 1000 bytes of flash even if `Ht16k33Module` is never used.
+        * Add `SimpleWireInterface`, a very simple software implementation of
+          I2C that supports sending only.
+        * Add `SimpleWireFastInterface`, same as `SimpleWireInterface` but
+          using one of the `digitalWriteFast` libraries.
+    * Improve `HardSpiInterface` and `HardSpiFastInterface`
+        * Fix Hardware SPI on ESP8266.
+            * Defer the initialization to `SPIClass::begin()` which knows to
+              call `pinMode(pin, SPECIAL)` instead of `pinMode(pin, OUTPUT)`
+              on the SCK and MOSI pins.
+            * Call `SPIClass::setHwCs(false)` to allow `HardSpiInterface` to
+              control the CS/SS pin, instead of being managed by the `SPIClass`.
+        * Add `SPIClass` as template parameter.
+        * Remove `dataPin` and `clockPin` parameters, since these pins
+          are implicit in the `SPIClass` object.
+        * Small reduction of flash (~30 bytes) and static memory (~2 bytes)
+          on AVR.
+    * Change order of `dioPin` and `clkPin` parameters in `SoftTmiInterface` and
+      `SoftTmiFastInterface` to match the order of `sda` and `scl` pins of
+      `TwoWire` and `SimpleWireInterface` classes.
+    * Rename `LedDisplay` to `PatternWriter` to remove extra layer of
+      abstraction.
+        * Pass `LedModule` into `Writer` classes (except `StringWriter` and
+          `StringScroller` which depend on `CharWriter`).
+    * Fix results of `MemoryBenchmark` for TeensyDuino so that the BASELINE
+      numbers include 3200 bytes of flash and 1100 bytes of static RAM from
+      pulling in `malloc()` and `free()`.
 * 0.6 (2021-05-26)
     * `HardSpiInterface`
         * Add support for microcontrollers with 2 SPI buses (ESP32 and STM32F1).